tools/managesf/requirements.txt
pecan
gerritlib
python-ldap
requests>=2.2.1
